8-1例を修正
